setAttribute: partial update with strings [fix #2651] (#2748)
* Enable the entity setAttribute API to accept partial updates of multi-property components when using a CSS-style property string
* Enable the use of the `clobber` flag under same scenario
* Remove single-use helper functions in a-entity for updating a single property of multi-property components
* Add unit tests and update a test where behavior changed

Details:

Previously, it was only possible to do partial updates when passing objects. Extending this behavior (and the clobber flag) to CSS-style strings requires differentiating those strings from simple property names, as a string in the second parameter could be either one.

This is handled by calling out to `utils.styleParser` and checking the return type, which is redundant to work done inside the component. This inefficiency could be reduced perhaps by reworking slightly the interface between the entity and component to pass the parsed result down; or a different mechanism of differentiation could be devised.

However, using strings in either capacity (CSS-style or the single property update form) is inherently a less efficient convenience for the object form, and their use should be discouraged when performance matters (e.g. `tick`).

The `clobber` flag (third parameter to setAttribute) behavior in this scenaro has now changed: previously, any time the second parameter was anything other than an object, `clobber` was treated internally as true. Now, CSS-style strings are treated the same as objects, where it defaults to false but is configurable by parameter.

This means that `el.setAttribute(‘myComponent’, ‘’)` previously would reset the component to its default values (because it was clobbering and not passing any new values). To get the same behavior as before, use `el.setAttribute(‘myComponent’, ‘’, true)`. One unit test was updated to reflect this change.

Finally, `entity.updateComponentProperty` has very little logic, is not a public API, and was only called by `setAttribute`; so it was simply merged in and its unit tests removed.