use Oculus Browser instead of Oculus Carmel
Use Oculus Browser instead of Oculus Carmel that is the old browser. I installed myself the Carmel Developer Preview app and found out later that it was actually an old version of the now called Oculus Browser, and Oculus Browser is installed by default with GearVR.
So the change is to avoid confusion.