Implement oculus-go-controls
This is mostly a copy and rename of the GearVR controls. This is to get everything boot strapped for now and then diverge them later while consolidating common functionality into a set of helpers.

The same for the tests, I've done a bit more consolidation in them, but also this is a copy.

For the controller we rely on the existing Gear VR controllers as well until we are able to upload Oculus Go specific controllers.