Remove references to discontinued sites (#4423)
As discussed here: https://github.com/aframevr/aframe/issues/4419 "The A-Frame Registry" and "Awesome A-Frame" are both now discontinued. I'm not sure if there's anything that should be added in their place (how are components selected for the Inspector these days, for example?) but the last word in the Issue was "just remove the references to the registry for now", so that's what I've done. (I've also inserted the alternatives that the registry's discontinuation readme suggests in its place.)