Update the TEXT component shader to conform GLSL 3 (#4656)
* Update TEXT shader to conform GLSL 3 ES version

The latest update of A-Frame to THREE r118 uses 'webgl2' context by default. That breaks the TEXT component, as it's shader should be updated to GLSL 3  #version 300 es.

* Update shader to conform GLSL #version 300 es 

The latest update of A-Frame to THREE r118 uses 'webgl2' context by default. That breaks the TEXT component, as it's shader should be updated to GLSL 3  #version 300 es.