Warning on html-and-primitives.md (#4860)
* Warning on html-and-primitives.md

aframe-physics-system does not work and has not worked with the newer version(s) of A-Frame for the past few months. Don McCurdy warned that they would stop maintaining frame-physics-system in [this blog post](https://www.donmccurdy.com/2020/09/06/projects-up-for-adoption/). Developers using this physics system are well aware of this bug and no official solution has been created yet. Possible solutions are being discussed [here](https://github.com/n5ro/aframe-physics-system/issues/187) but since it has been unusable for quite a while now, I think it would be reasonable to warn people that it may not work properly.

* Update wording and links

Co-authored-by: Diego Marcos <diego.marcos@gmail.com>