Update text plane size following text mutation (fixes #2764). (#5357)
The plane width/height was being calculated when the text value was first set,
but would not update when the text changed.

Documentation has also been updated to advise using `0` rather than `auto`
for autoscaling behavior.  This is based on the the comment
[here](https://github.com/aframevr/aframe/issues/2837#issuecomment-315923754)
suggesting that `auto` is not a valid value for geometry (NaN is falsy, so it
happened to still work).